+++ test.c
@@ -15,8 +15,8 @@ void list_add_first_test()
 {
     linked_list_int_t **list;
     
-    int value1 = 1;
-    int value2 = 2;
+    int64_t value1 = 1;
+    int64_t value2 = 2;
     
     
     list = linked_list_int_create();
@@ -44,8 +44,8 @@ void list_add_last_test()
 {
     linked_list_int_t **list;
     
-    int value1 = 1;
-    int value2 = 2;
+    int64_t value1 = 1;
+    int64_t value2 = 2;
     
     
     list = linked_list_int_create();
